backend/requirements.txt
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Groq API
groq>=0.4.0

# Image Processing
pillow>=10.0.0

# Environment
python-dotenv>=1.0.0

# CORS
